First menu version (#9)
* First mockup UI

* Camera test scene update

* UI updates and corrections

* Unity scene
