+++ Assets/InitialMenuUI.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8903728edd3343f4f9488e1fb533a06f
+guid: ed89d5cf05e55474fa279fd785ac048f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
