+++ Assets/Scripts/UI/IngameMenuUI.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8903728edd3343f4f9488e1fb533a06f
+guid: a71a9cd879ce5e348bb4e3fc7d1544f9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
